This is a great place to start building the app
At this point the app has the bare basic functionality we are looking for. This is a good foundation. From here we can expand the app.
I have added a feature so that a team can be signed up when they are trying to sign in(in case they failed to sign up in the first place).